Fix ID type in Game API
- changed type in game repo to Long, which is standard id type for sql based dbs

+++ api/src/main/java/org/hungerhome/soccermanager/api/repository/GameRepository.java
@@ -3,4 +3,4 @@ package org.hungerhome.soccermanager.api.repository;
 import org.hungerhome.soccermanager.api.entity.Game;
 import org.springframework.data.repository.CrudRepository;
 
-public interface GameRepository extends CrudRepository<Game, Integer> { }
+public interface GameRepository extends CrudRepository<Game, Long> { }
